TempHumidityDataLog/requirements.txt
pyserial==3.5
gspread==5.9.0
oauth2client==4.1.3
